pyproject.toml
[project]
name = "mcp-serial-bridge"
version = "0.1.0"
description = "MCP server for controlling serial devices via AI agents"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "46nori" }]
requires-python = ">=3.11"
keywords = ["mcp", "serial", "pyserial", "ai-agent", "model-context-protocol"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Communications",
    "Topic :: System :: Hardware",
]
dependencies = [
    "mcp[cli]>=1.9.0",
    "pyserial>=3.5",
]

[project.urls]
Repository = "https://github.com/46nori/mcp-serial-bridge"
